fix(security): use int instead of long for curDigit in parseJiaoFen
Fixes CodeQL alerts #3 and #4 (java/implicit-cast-in-compound-assignment).

curDigit only ever holds a single Chinese digit (0-9), so int is the correct type; long unnecessarily forced implicit narrowing in the `fen += curDigit * 10` and `fen += curDigit` compound assignments.

+++ mica-ppocr-structured/src/main/java/net/dreamlu/mica/ai/ppocr/structured/parser/invoice/UpperMoneyConverter.java
@@ -191,7 +191,7 @@ final class UpperMoneyConverter {
 	 */
 	private static int parseJiaoFen(String s, int start) {
 		int fen = 0;
-		long curDigit = 0;
+		int curDigit = 0;
 		boolean sawJiao = false;
 		boolean sawFen = false;
 		for (int i = start; i < s.length(); i++) {
